Close works, and movement
The escape will now properly close the game, and you can now move with
wasd. The right arrow currently rotates the mc, and the other keys need
to be programmed as well for rotation. Joystick will be added next, and
if I have enough time the fox will have a walking animation.